backend/requirements.txt
# Backend Python Dependencies for Policy RAG Application
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Vector Database
pinecone-client==3.0.2

# LangChain & LangGraph
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-text-splitters==0.0.1
langgraph==0.2.16

# Document Processing
pypdf==4.0.1
pdfplumber==0.10.4

# Embeddings & LLMs
openai==1.10.0
anthropic==0.18.1

# HTTP client for Ollama
requests==2.31.0
httpx==0.26.0

# Multimodal / Vision
Pillow==10.2.0
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Caching
redis==5.0.1
hiredis==2.3.2  # Performance boost for Redis
